Level 3 Puzzle Started
Started work on the puzzle in the main room of Level 3. It will have the player moving around the room attempting to place boxes on 3 pressure plates located around the room, each one opening a door leading to the end and only after all 3 are open will the player be able to reach the end.

Created the first part where the player will have to shoot a fireball between some bars in order to blow up a barrel and open up a wall. This will allow the player to reach some boxes that they will be able to use to reach the top of a wall and the first pressure plate.

- Changed some names on objects to better understand what objects are
- Added puzzles to main room (See above)
- Moved portal folder into prefabs