Level 5 1st Puzzle
- Finished the layout
- Added doors (will change at a later date)
- Began work on first puzzle

The objective of this level will be to work your way around the building, trying to get into the courtyard in the centre. The player will need to solve various puzzles in all the rooms to reach it. The first puzzle will require all 3 spells, needing the player to burn down a barricade then make a plank grow before picking it up and placing it down in order to walk across it like a ramp.